requirements.txt
requests
beautifulsoup4
langchain-community
langchain-core
langchain-groq
langgraph
qdrant-client
python-dotenv
kiwipiepy
rank_bm25
httpx
groq
fastapi
uvicorn
pydantic
easyocr
opencv-python-headless
mmh3